edit functionality add

+++ src/main/java/com/example/phonebook/repositories/EmployeeRep.java
@@ -4,6 +4,8 @@ import com.example.phonebook.domain.Employee;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
-public interface EmployeeRep extends CrudRepository<Employee, Long> {
+import java.util.Optional;
+
+public interface EmployeeRep extends CrudRepository<Employee, Integer> {
     List<Employee> findByFirstNameLike(String firstName);
 }
